requirements.txt
pyrogram==2.0.106
requests==2.31.0
lexica-api==1.5.4
tgcrypto==1.2.5
asyncio==3.4.3
python-dotenv==1.0.0
openai==0.28.0
gtts
aiohttp== 3.9.3
aiofiles
yt_dlp
youtube-search==2.1.2
youtube-search-python==1.6.6
pydub
wget
ffmpeg-python==0.2.0
beautifulsoup4
SafoneAPI
Abg==2.3.1
pymongo
motor
dnspython
pillow
numpy
gunicorn
flask

